+++ src/index.js
@@ -5,3 +5,4 @@ export { default as SentFeedback } from "./components/SentFeedback";
 export { default as Table } from "./components/Table";
 export { default as Balance } from "./components/Balance";
 export { default as Button } from "./components/Button";
+export { default as Label } from "./components/Label";